Add optional parameters `direction` and `stranded` to `get_bps`.
There is a rotational symmetry pattern on the heatmaps, which is caused by the fact
that the windows generated from the start positions point to the inside of the aligned
region, and the windows generated from the end positions point to the outside.  By
marking them as minus strand, the sequences are flipped and all orientations are the
same

After correcting this, there is a mirror symmetry that remains.  When the aligned
regions are narrower than the windows, plotting windows from both sides causes the
same sequences to be plotted twice.  The "direction" option allows to select only
the start or the end of the alignments for defining windows.