Store the intermediate results in the gr_ob object directly.
The main reason is that some of the tests resuts such as
gr_ob$con_met_total are logical-Rle, and therefore can not be used
to subset plain vectors (which do not accept S4 objects as indexes)
However, they can be used to subset GRanges.

Additional advantages are that when debugging it is easier
to inspect all the values side to side.