Plot a three-genome comparison.
The plot looks nice, but it took a lot of time to find a
region that can be plotted.  Thus, it does not refect
the data.

One problem is that the O and N assemblies are too
fragmented, thus blocks that are on the same chromosome
end up on different contigs, and can not be plotted easily
with that approach