Use Bioconductor's `slidingWindows` function.
The sequence information of a GRanges/GBreaks object can easily be converted
to a GRanges object representing the chromosomes, which is passed to the
`slidingWidnows` function.  It returns a `GRangesList` which is converted back
to `GRanges` with the `unlist()` function.  Compared with the `makeSlidingWindows()`
function, it may add an incomplete tile at the end, so the next line removes
it if needed.